Completed game parser

+++ src/main/java/de/ailis/wlandsuite/game/parts/CentralDirectory.java
@@ -229,7 +229,7 @@ public class CentralDirectory
      *            The offset
      */
 
-    public void setActionClassOffset(byte actionClass, int offset)
+    public void setActionClassOffset(int actionClass, int offset)
     {
         this.actionClassMasterTable[actionClass] = offset;
     }
